Extract install spec deployment support
